+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changes
 
+## 1.x.x [2022/XX/XX]
+
+**New Features:**
+
+**Enhancements:**
+
+- Changed Event ID Statistics option to Event ID Metrics option. (`-s, --statistics`  -> `-M, --metrics`) (#706) (@hitenkoku)
+  (Note: `statistics_event_info.txt` was changed to `event_id_info.txt`.)
+
+**Bug Fixes:**
+
 ## v1.6.0 [2022/09/16]
 
 **New Features:**